MoreCast v2 front-end (#2657)
Note that this is not yet wired up to the API, there is some sample data in place for the HRDPS model. The is data for one weather station in Cariboo Fire Centre and one in the Southeast Fire Centre.

Selecting a different model will show no data. The checkboxes in the Station panel are wired up and should toggle data in the data grid. 

Selected fire centre and weather model are saved in local storage.

The datagrid is not yet editable (that's an upcoming ticket!).

I've disabled a lot of functionality including sorting/filtering/reordering, except for the Station column.

More tests are on the way!