Round numbers when editing (#2699)
This PR fixes a couple of UI quirks:
1. Rounds numbers to a specific decimal place when editing (model data has many decimal places, but we want 0 or one depending on the column
2. It was possible to encounter a situation where a user selects the same start/end date, menaing we should show data for that date, yet the data grid was blank (fix was the comparison in MoreCast2Page.tsx).

Finally, React was throwing an error when attempting to edit a cell with a N/A value. The reason was that the underlying NaN was being passed to an `<input type=number>` element which can't handle 'NaN'. Instead, we now pass in a stringified 'NaN' which gives us a console warning, but not an error.