Add checks for missing stations (#2838)
The pdf endpoint assumes the list of stations from WF1 is the same list of stations for a HFI request. This is not true since we made the change to render the HFI table based on the request.

The calculations are based on the HFI request as well so now every function/artifact is based on it.