Morecast: Validation on editable fields (#3919)
Adds an optional `validator` function for `IndeterminateField` that gets called in `preProcessEditCellProps` callback for cells in edit mode to set the error prop if validation doesn't pass for the forecast parameter. The same `validator` function is passed down to cells in view mode to do the same thing.

Adds the `EditInputCell` component as the render component for cells in edit mode that gets rendered by the `renderEditCell` callback on the `GridColDef`. `EditInputCell` props include an empty or non-empty `error` string that drives the visibility of the tooltip and red border.

Similar behaviour for the `renderCell` callback for cells in view mode, except it doesn't receive the `error` string in props so it calls `validator` itself. This is to keep the same behaviour whether the user is actively editing or not and the current value is invalid.

Opted to prevent edits from committing locally when invalid so that invalid values aren't stored in drafts.