SFMS/NATS timezone issue (#4309)
- Fixes timezone issue in our SFMS/HFI/NATS processing pipeline
- SFMS data is stored in s3 with a date in America/Vancouver time. We store run_datetime in our db in UTC.
If we receive data after a certain time (ie 5:00pm PST/PDT), that will be after midnight in UTC time. If we used that
UTC datetime as a date to find data stored in s3, it would give us the incorrect date. This was happening in a couple places in that pipeline