ASA: Store and Precalculate Fuel Area Stats for Each Fuel Grid (#4428)
Co-authored-by: dgboss <darren.boss@gov.bc.ca>

Adds a job that can be triggered to look up the `FUEL_RASTER_NAME` as the raster object name and expected content hash respectively in the environment. Checks the content hash and copies the raster object to `sfms/static/fuel/yyyy/fbpyyyy_vx.tif` where `yyyy` is the current year, x is the latest version for that year.
Stores the year, version, object key, resolution for each processed raster in the `fuel_type_raster` table.

- Moves `raster_addresser`, `rdps_filename_marshaller`, `dataset_common` to `wps_shared`